requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
cryptography>=41.0.0
argon2-cffi>=23.1.0
watchdog>=3.0.0
pyinstaller>=5.0.0